requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
openai>=1.40.0
werkzeug>=3.0.0